refactor: move useAtomInput to custom hooks #37
